Create an agent for integral-free RL
The weights converge to some value but not the optimal.
This might be due to the persistent excitation.
To converge the weights, a data stack like concurrent learning should be
used.

The stability is only guaranteed when the system is stable.
It is necessary to invent a method to remedy this problem.